+++ changelog.md
@@ -1,3 +1,6 @@
+### 0.0.2 30/12/2021
+ -update system generate custom link
+ -update system generate random link
 ### 0.0.1 26/12/2021
  -update README.md
  -initial commit